# internal/handlers/http/v1/openapi/openapi.yaml
openapi: "3.1.0"
info:
  title: Itmo Calendar
  description: Itmo calendar sync service
  version: "1.0.0"

servers:
  - url: /api/v1

security: []

components:
  securitySchemes:
    JWT:
      type: apiKey
      in: header
      name: X-Auth-Token
      description: JWT token for user authentication

  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Service Unavailable"
        message:
          type: string
          example: "The service is currently unavailable. Please try again later."

    SubscribeRequest:
      type: object
      required:
        - isu
        - password
      properties:
        isu:
          type: integer
          format: int64
          example: 123456789
        password:
          type: string
          example: "user_password"

    SubscribeResponse:
      type: object
      properties:
        message:
          type: string
          example: "Subscription successful. iCal generated."

    Lesson:
      type: object
      required:
        - subject
        - type
        - teacher_name
        - room
        - building
        - format
        - group
        - time_start
        - time_end
      properties:
        subject:
          type: string
          example: "Mathematics"
        type:
          type: string
          example: "Lecture"
        teacher_name:
          type: string
          example: "Dr. Ivanov"
        room:
          type: string
          example: "101"
        note:
          type: string
          example: "Bring calculator"
        building:
          type: string
          example: "Main"
        format:
          type: string
          example: "Offline"
        group:
          type: string
          example: "A1"
        zoom_url:
          type: string
          example: "https://zoom.us/j/123456789"
        time_start:
          type: string
          format: date-time
          example: "2024-06-01T09:00:00Z"
        time_end:
          type: string
          format: date-time
          example: "2024-06-01T10:30:00Z"

    ScheduleItem:
      type: object
      required:
        - date
        - lessons
      properties:
        date:
          type: string
          format: date
          example: "2024-06-01"
        lessons:
          type: array
          items:
            $ref: "#/components/schemas/Lesson"

paths:
  /health:
    get:
      summary: Health check endpoint.
      operationId: healthCheck
      security: []
      description: Verifies the API is operational and returns its status.
      tags:
        - System
      responses:
        "200":
          description: Service is healthy.
        default:
          description: Error response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /{isu}/schedule:
    get:
      summary: Get user's schedule by ISU.
      operationId: getSchedule
      description: Returns the schedule for the user with the given ISU.
      tags:
        - Schedule
      parameters:
        - name: isu
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ISU of the user.
      responses:
        "200":
          description: User's schedule.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScheduleItem"
        default:
          description: Error response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /{isu}/ical:
    get:
      summary: Get user's iCal file by ISU.
      operationId: getICal
      description: Returns the iCalendar (.ics) file for the user with the given ISU.
      tags:
        - CalDav
      parameters:
        - name: isu
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ISU of the user.
      responses:
        "200":
          description: iCal file.
          content:
            text/calendar:
              schema:
                type: string
        default:
          description: Error response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /subscribe:
    post:
      summary: Subscribe and generate iCal for user.
      operationId: subscribeSchedule
      description: Subscribes user by ISU and password, generates and stores iCal file.
      tags:
        - CalDav
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscribeRequest"
      responses:
        "200":
          description: Subscription successful.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscribeResponse"
        default:
          description: Error response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"